Generic Rest Client
